test(query-builder): split query-builder-joins.test.ts into basic, composite-pk, and take/with-joins
