+++ test/functional/query-builder/join/basic/entity/Post.ts
@@ -1,11 +1,11 @@
-import { Entity } from "../../../../../src/decorator/entity/Entity"
-import { PrimaryGeneratedColumn } from "../../../../../src/decorator/columns/PrimaryGeneratedColumn"
-import { Column } from "../../../../../src/decorator/columns/Column"
-import { ManyToOne } from "../../../../../src/decorator/relations/ManyToOne"
-import { ManyToMany } from "../../../../../src/decorator/relations/ManyToMany"
-import { JoinTable } from "../../../../../src/decorator/relations/JoinTable"
-import { OneToOne } from "../../../../../src/decorator/relations/OneToOne"
-import { JoinColumn } from "../../../../../src/decorator/relations/JoinColumn"
+import { Entity } from "../../../../../../src/decorator/entity/Entity"
+import { PrimaryGeneratedColumn } from "../../../../../../src/decorator/columns/PrimaryGeneratedColumn"
+import { Column } from "../../../../../../src/decorator/columns/Column"
+import { ManyToOne } from "../../../../../../src/decorator/relations/ManyToOne"
+import { ManyToMany } from "../../../../../../src/decorator/relations/ManyToMany"
+import { JoinTable } from "../../../../../../src/decorator/relations/JoinTable"
+import { OneToOne } from "../../../../../../src/decorator/relations/OneToOne"
+import { JoinColumn } from "../../../../../../src/decorator/relations/JoinColumn"
 import { User } from "./User"
 import { Category } from "./Category"
 import { CategoryWithCompositePK } from "./CategoryWithCompositePK"
