+++ test/functional/query-builder/join/basic/entity/User.ts
@@ -3,7 +3,7 @@ import {
     Entity,
     OneToMany,
     PrimaryGeneratedColumn,
-} from "../../../../../src"
+} from "../../../../../../src"
 import { Photo } from "./Photo"
 
 @Entity()
