+++ test/functional/query-builder/join/composite-pk/entity/Image.ts
@@ -1,0 +1,15 @@
+import { Entity } from "../../../../../../src/decorator/entity/Entity"
+import { PrimaryGeneratedColumn } from "../../../../../../src/decorator/columns/PrimaryGeneratedColumn"
+import { Column } from "../../../../../../src/decorator/columns/Column"
+
+@Entity()
+export class Image {
+    @PrimaryGeneratedColumn()
+    id: number
+
+    @Column()
+    name: string
+
+    @Column()
+    isRemoved: boolean = false
+}
