+++ test/functional/query-builder/join/composite-pk/entity/User.ts
@@ -1,0 +1,19 @@
+import {
+    Column,
+    Entity,
+    OneToMany,
+    PrimaryGeneratedColumn,
+} from "../../../../../../src"
+import { Photo } from "./Photo"
+
+@Entity()
+export class User {
+    @PrimaryGeneratedColumn()
+    id: number
+
+    @Column()
+    name: string
+
+    @OneToMany(() => Photo, (photo) => photo.user)
+    photos: Photo[]
+}
