--- test/functional/query-builder/join/entity/Tag.ts
@@ -1,12 +1,0 @@
-import { Entity } from "../../../../../src/decorator/entity/Entity"
-import { PrimaryGeneratedColumn } from "../../../../../src/decorator/columns/PrimaryGeneratedColumn"
-import { Column } from "../../../../../src/decorator/columns/Column"
-
-@Entity()
-export class Tag {
-    @PrimaryGeneratedColumn()
-    id: number
-
-    @Column()
-    name: string
-}
