+++ test/functional/query-builder/take/with-joins/entity/Photo.ts
@@ -1,0 +1,19 @@
+import {
+    Column,
+    Entity,
+    ManyToOne,
+    PrimaryGeneratedColumn,
+} from "../../../../../../src"
+import { User } from "./User"
+
+@Entity()
+export class Photo {
+    @PrimaryGeneratedColumn()
+    id: number
+
+    @Column()
+    name: string
+
+    @ManyToOne(() => User, (user) => user.photos)
+    user: User
+}
